// src/app/utils/interfaces/group.ts
// Generated by https://quicktype.io

export interface Group {
    _metadata?: Metadata;
    isActive?: boolean;
    users?: any[];
    _id?: string;
    name?: string;
    description?: string;
    app?: string;
    roles?: any[];
    __v?: number;
}

export interface Metadata {
    version?: Version;
    deleted?: boolean;
    lastUpdated?: string;
    createdAt?: string;
}

export interface Version {
    document?: number;
}